+++ src/hiero_sdk_python/__init__.py
@@ -52,6 +52,7 @@ from .tokens.token_unpause_transaction import TokenUnpauseTransaction
 from .tokens.token_pause_transaction import TokenPauseTransaction
 from .tokens.token_airdrop_claim import TokenClaimAirdropTransaction
 from .tokens.assessed_custom_fee import AssessedCustomFee
+from .tokens.token_association import TokenAssociation
 
 # Transaction
 from .transaction.transaction import Transaction
@@ -204,6 +205,7 @@ __all__ = [
     "TokenPauseTransaction",
     "TokenUnpauseTransaction",
     "AssessedCustomFee",
+    "TokenAssociation",
 
     # Transaction
     "Transaction",