+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-﻿# Changelog
+﻿﻿# Changelog
 
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org).
@@ -7,7 +7,8 @@ This changelog is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.
 ## [Unreleased]
 
 ### Src
-- 
+
+- Exposed all missing fields from `TransactionRecord` protobuf (#1636) 
 
 ### Examples